put ad code into separate file, enhance error checking
